requirements.txt
Pillow==4.0.0
django-jet==1.0.6
django-modeltranslation
solid_i18n